NEW - bug 470818: General cleanup of testopia classes https://bugzilla.redhat.com/show_bug.cgi?id=470818
Detect version number, use plan with that version number (must be created manually, but all testcases will be created in it automagically).

Getting the version number is up to the product automation implementation.  It can be set by the static method TestopiaTestNGListener.setVersion(String).  This should be done in the BeforeSuite method(s) - before the test actually starts

git-svn-id: svn://svn.rhndev.redhat.com:55555/svn/rhn-svn/trunk/qa/automation/webui-framework@178824 ad8b2fa6-1fe3-0310-8810-ca5544d99bb5